// src/api/types.d.ts
export interface paths {
// ...
  '/3/movie/{movie_id}': {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /**
     * Details
     * @description Get the top level details of a movie by ID.
     */
    get: operations['movie-details']
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
// ...
  '/3/movie/{movie_id}/images': {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /**
     * Images
     * @description Get the images that belong to a movie.
     */
    get: operations['movie-images']
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
// ...
  '/3/movie/{movie_id}/credits': {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /** Credits */
    get: operations['movie-credits']
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
// ...
  '/3/movie/{movie_id}/similar': {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /**
     * Similar
     * @description Get the similar movies based on genres and keywords.
     */
    get: operations['movie-similar']
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
// ...
  '/3/movie/{movie_id}/videos': {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /** Videos */
    get: operations['movie-videos']
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
// ...
  '/3/movie/upcoming': {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /**
     * Upcoming
     * @description Get a list of movies that are being released soon.
     */
    get: operations['movie-upcoming-list']
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
  '/3/movie/now_playing': {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /**
     * Now Playing
     * @description Get a list of movies that are currently in theatres.
     */
    get: operations['movie-now-playing-list']
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
// ...
}
export type webhooks = Record<string, never>
export interface components {
  schemas: never
  responses: never
  parameters: never
  requestBodies: never
  headers: never
  pathItems: never
}
export type $defs = Record<string, never>
export interface operations {
  'search-movie': {
    parameters: {
      query: {
        query: string
        include_adult?: boolean
        language?: string
        primary_release_year?: string
        page?: number
        region?: string
        year?: string
      }
      header?: never
      path?: never
      cookie?: never
    }
    requestBody?: never
    responses: {
      /** @description 200 */
      200: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/json': {
            /**
             * @default 0
             * @example 1
             */
            page: number
            results?: {
              /**
               * @default true
               * @example false
               */
              adult: boolean
              /** @example /hZkgoQYus5vegHoetLkCJzb17zJ.jpg */
              backdrop_path?: string
              genre_ids?: number[]
              /**
               * @default 0
               * @example 550
               */
              id: number
              /** @example en */
              original_language?: string
              /** @example Fight Club */
              original_title?: string
              /** @example A ticking-time-bomb insomniac and a slippery soap salesman channel primal male aggression into a shocking new form of therapy. Their concept catches on, with underground "fight clubs" forming in every town, until an eccentric gets in the way and ignites an out-of-control spiral toward oblivion. */
              overview?: string
              /**
               * @default 0
               * @example 73.433
               */
              popularity: number
              /** @example /pB8BM7pdSp6B6Ih7QZ4DrQ3PmJK.jpg */
              poster_path?: string
              /** @example 1999-10-15 */
              release_date?: string
              /** @example Fight Club */
              title?: string
              /**
               * @default true
               * @example false
               */
              video: boolean
              /**
               * @default 0
               * @example 8.433
               */
              vote_average: number
              /**
               * @default 0
               * @example 26279
               */
              vote_count: number
            }[]
            /**
             * @default 0
             * @example 2
             */
            total_pages: number
            /**
             * @default 0
             * @example 39
             */
            total_results: number
          }
        }
      }
    }
  }
// ...
}